+++ pct_scalewiz/components/project_report.py
@@ -21,21 +21,25 @@ class ProjectReport(ttk.Frame):
 
         parent.render(lbl, ent, 0)
 
+        # matplotlib stuff
         # todo implement color selection
         # colorsLbl = ttk.Label(self, text="Plot color cycle:")
         # colorsEnt = ttk.Entry(self)
         # parent.render(colorsLbl, colorsEnt, 1)
+
         # todo implement legend options
         # legendLbl = ttk.Label(self, text="Legend location:")
         # legendEnt = ttk.Combobox(self)
         # parent.render(legendLbl, legendEnt, 2)
 
-    def loadTemplate(self, *args):
-        path = filedialog.askopenfilename(
-            initialdir='C:"',
-            title="Select report template:",
-            filetypes=[("Excel files", "*.xlsx")],
-        )
+    # support for excel exports was un-implemented in favor of a more general approach
 
-        if not (path == ""):
-            self.parent.editorProject.template.set(path)
+    # def loadTemplate(self, *args):
+    #     path = filedialog.askopenfilename(
+    #         initialdir='C:"',
+    #         title="Select report template:",
+    #         filetypes=[("Excel files", "*.xlsx")],
+    #     )
+
+    # if not (path != ""):
+    #     self.parent.editorProject.template.set(path)