+++ CHANGELOG.md
@@ -13,6 +13,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
  - rinse windows will be titled after the system they control
  - requests for rinse cycles will be ignored if a test is already running
  - the Live Plot stops updating (clearing itself) at the end of a test
+### Fixed
+ - observed baseline pressure reporting
 ### Removed
  - dependency on openpyxl
  - dependency on PIL